improve build pipeline by isolating stages
  * introduce build "workspaces"
    - workspaces are created within stage build directories
      to organize generated code
      -- generated code is now tied to a build as it should be
    - each stage (workspace) gets own copy of CMakeLists.txt
      and can thus properly distinguish when and when not files
      are changed
  * stage 1 data reader now "patches" matchable headers in place
    instead of making copies
  * remove STAGE_X compile definitions making code stage agnostic
  * fixes "retain" build features (incremental building)
    - per stage CMakeLists.txt make this easy
  * introduce "retain_leaves" build feature
    - allows quick building when changing anything but "leaf"
      templates
  * stage 0 build directory now parallel to main build directory,
    instead of nested under it (suffixed with '_stage_0')

+++ scripts/build_and_install.py
@@ -24,11 +24,15 @@ def usage():
     print('                                    * use system compiler by default\n')
     print('    -r  --retain                  retain previous build')
     print('                                    * continuation/incremental building')
-    print('                                    * implies \'-m\'\n')
+    print('                                    * implies \'-l\'\n')
+    print('    -l  --retain_leaves           retain most of previous build')
+    print('                                    * implies \'-m\'')
+    print('                                    * parents are rebuilt')
+    print('                                    * matchmaker interface is rebuilt')
+    print('                                    * leaves are left alone\n')
     print('    -m  --retain_matchables       retain generated matchables')
-    print('                                    * skips build matchable step')
-    print('                                    * skips build data_reader step')
-    print('                                    * skips run data_reader step\n')
+    print('                                    * skips building of matchable')
+    print('                                    * skips building and running of data readers\n')
     print('    -j  --jobs                    max jobs')
     print('                                    * default is cpu count [' +                                 \
             str(multiprocessing.cpu_count()) + '])\n')
@@ -44,8 +48,8 @@ def usage():
 
 
 
-def build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, install_dir, atomic_libs,
-                      q, debug):
+def build_and_install(use_clang, retain, retain_leaves, retain_matchables, jobs, build_dir, install_dir,
+                      atomic_libs, q, debug):
     print('\n\n************** matchmaker **************\n\n')
     print('starting stage 0 ...\n\n')
 
@@ -57,96 +61,28 @@ def build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, ins
     suffix = ''
     if q:
         suffix = '_q'
-    elif atomic_libs:
-        suffix = '_atomic'
+    if atomic_libs:
+        suffix = suffix + '_atomic'
 
-    if not retain:
-        print('generate matchmaker code...\n')
-        prepare_letters_cmd = ['scripts/prepare_letters.py']
-        if q:
-            prepare_letters_cmd.append('-q')
-        if atomic_libs:
-            prepare_letters_cmd.append('-a')
-        if subprocess.run(prepare_letters_cmd).returncode != 0:
-            print('prepare_letters.py failed')
-            exit(1)
-        print('\n\nmatchmaker code ready!')
-
-        if not retain_matchables:
-
-            # first build and install matchable
-            print('\n\nbuild and install matchable:\n\n')
-            build_matchable_cmd = ['matchable/scripts/build_and_install.py']
-            build_matchable_cmd.append('-j' + jobs)
-            build_matchable_cmd.append('-l')
-            if use_clang:
-                build_matchable_cmd.append('-c')
-
-            build_matchable_cmd.append('-b')
-            build_matchable_cmd.append('build' + suffix)
-            build_matchable_cmd.append('-i')
-            build_matchable_cmd.append('../install' + suffix)
-
-            if subprocess.run(build_matchable_cmd).returncode != 0:
-                print('matchable failed to build')
-                exit(1)
-            print('\n\nmatchable ready!')
-
-            # then build and install data reader for stage 0
-            print('\n\nbuild and install data reader for stage 0:\n\n')
-            build_data_reader_cmd = ['data_reader/stage_0/scripts/build_and_install.py']
-            build_data_reader_cmd.append('-b')
-            build_data_reader_cmd.append('build' + suffix)
-            build_data_reader_cmd.append('-i')
-            build_data_reader_cmd.append('../install' + suffix)
-            build_data_reader_cmd.append('-m')
-            build_data_reader_cmd.append('../../matchable/install' + suffix)
-            if use_clang:
-                build_data_reader_cmd.append('-c')
-            if subprocess.run(build_data_reader_cmd).returncode != 0:
-                print('data reader failed to build')
-                exit(1)
-            print('\n\nstage 0 data reader ready!')
-
-            # prepare_matchables for stage 0
-            print('\n\ncreate stage 0 matchables...')
-            prepare_matchables_cmd = ['scripts/prepare_matchables.py']
-            prepare_matchables_cmd.append('-l')
-            prepare_matchables_cmd.append('data_reader/stage_0/install' + suffix)
-            if q:
-                prepare_matchables_cmd.append('-q')
-            if atomic_libs:
-                prepare_matchables_cmd.append('-a')
-            if subprocess.run(prepare_matchables_cmd).returncode != 0:
-                print('failed to prepare matchables for stage 0')
-                exit(1)
-            print('\n\nstage 0 matchables ready!')
-
-            # create longest_word_stage_0.h
-            stage_0_longest_word_file = matchmaker_root + 'generated_include' + suffix                     \
-                    + '/matchmaker/longest_word_stage_0.h'
-            with open(stage_0_longest_word_file, 'w') as f:
-                f.write('#pragma once\ninline int const LONGEST_WORD{0};\n')
-
-
-    # note: stage 0 builds and installs within the build dir
-
+    # vars for build and install directories
     if build_dir == '':
         build_dir = matchmaker_root + '/build'
     while build_dir[-1] == '/':
         build_dir = build_dir[:-1]
-    build_dir = build_dir + suffix + '/'
-    stage_0_build_dir = build_dir + 'build_stage_0/'
+    build_dir = build_dir + suffix
+    stage_0_build_dir = build_dir + '_stage_0/'
+    build_dir = build_dir + '/'
 
     if install_dir == '':
         install_dir = matchmaker_root + '/install'
     while install_dir[-1] == '/':
         install_dir = install_dir[:-1]
     install_dir = install_dir + suffix + '/'
-    stage_0_install_dir = build_dir + 'install_stage_0/'
+    stage_0_install_dir = stage_0_build_dir + 'install/'
 
-    if not retain:
+    if not retain and not retain_leaves:
         shutil.rmtree(install_dir, ignore_errors=True)
+        shutil.rmtree(stage_0_build_dir, ignore_errors=True)
         shutil.rmtree(build_dir, ignore_errors=True)
     if not os.path.exists(install_dir):
         os.makedirs(install_dir)
@@ -157,18 +93,100 @@ def build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, ins
     if not os.path.exists(stage_0_install_dir):
         os.makedirs(stage_0_install_dir)
 
+    stage_0_workspace_dir = stage_0_build_dir + 'workspace/'
+
+    if not retain or not os.path.exists(stage_0_workspace_dir):
+
+        # create stage 0 workspace as needed
+        if not retain_leaves:
+            shutil.rmtree(stage_0_workspace_dir, ignore_errors=True)
+        if not os.path.exists(stage_0_workspace_dir):
+            os.makedirs(stage_0_workspace_dir)
+            shutil.copy(matchmaker_root + 'workspace_common/CMakeLists.txt',
+                        stage_0_workspace_dir + 'CMakeLists.txt')
+            shutil.copy(matchmaker_root + 'workspace_common/config.cmake.in',
+                        stage_0_workspace_dir + 'config.cmake.in')
+
+        print('preparing matchmaker code...\n')
+        prepare_letters_cmd = ['scripts/prepare_letters.py', '-w', stage_0_workspace_dir]
+        if q:
+            prepare_letters_cmd.append('-q')
+        if retain_leaves:
+            prepare_letters_cmd.append('-p')
+        if subprocess.run(prepare_letters_cmd).returncode != 0:
+            print('prepare_letters.py failed')
+            exit(1)
+        print('\n\nmatchmaker code ready!')
+
+        if not retain_matchables and not retain_leaves:
+
+            # first build and install matchable
+            print('\n\nbuild and install matchable:\n\n')
+            build_matchable_cmd = ['matchable/scripts/build_and_install.py']
+            build_matchable_cmd.append('-j' + jobs)
+            build_matchable_cmd.append('-l')
+            if use_clang:
+                build_matchable_cmd.append('-c')
+
+            build_matchable_cmd.append('-b')
+            build_matchable_cmd.append('build_mm' + suffix)
+            build_matchable_cmd.append('-i')
+            build_matchable_cmd.append('../install_mm' + suffix)
+
+            if subprocess.run(build_matchable_cmd).returncode != 0:
+                print('matchable failed to build')
+                exit(1)
+            print('\n\nmatchable ready!')
+
+            # then build and install data reader for stage 0
+            print('\n\nbuild and install data reader for stage 0:\n\n')
+            build_data_reader_cmd = ['data_reader/stage_0/scripts/build_and_install.py']
+            build_data_reader_cmd.append('-b')
+            build_data_reader_cmd.append('build_mm' + suffix)
+            build_data_reader_cmd.append('-i')
+            build_data_reader_cmd.append('../install_mm' + suffix)
+            build_data_reader_cmd.append('-m')
+            build_data_reader_cmd.append('../../matchable/install_mm' + suffix)
+            if use_clang:
+                build_data_reader_cmd.append('-c')
+            if subprocess.run(build_data_reader_cmd).returncode != 0:
+                print('data reader failed to build')
+                exit(1)
+            print('\n\nstage 0 data reader ready!')
+
+            # prepare_matchables for stage 0
+            print('\n\ncreating stage 0 matchables...')
+            prepare_matchables_cmd = ['scripts/prepare_matchables.py', '-w', stage_0_workspace_dir]
+            prepare_matchables_cmd.append('-d')
+            prepare_matchables_cmd.append('data_reader/stage_0/install_mm' + suffix)
+            if q:
+                prepare_matchables_cmd.append('-q')
+            if subprocess.run(prepare_matchables_cmd).returncode != 0:
+                print('failed to prepare matchables for stage 0')
+                exit(1)
+            print('\n\nstage 0 matchables ready!')
+
+            # create longest_word_stage_0.h
+            stage_0_longest_word_file = stage_0_workspace_dir + 'generated_include'                        \
+                    + '/matchmaker/longest_word.h'
+            with open(stage_0_longest_word_file, 'w') as f:
+                f.write('#pragma once\ninline int const LONGEST_WORD{0};\n')
+
+
+
+
     os.chdir(stage_0_build_dir)
 
     print('\n\nMATCHMAKER!\n\n')
 
     cmake_cmd = ['cmake', '-DCMAKE_INSTALL_PREFIX=' + stage_0_install_dir]
 
-    cmake_cmd.append('-Dstage=0')
-
     if debug:
         cmake_cmd.append('-DCMAKE_BUILD_TYPE=Debug')
     else:
         cmake_cmd.append('-DCMAKE_BUILD_TYPE=Release')
+
+    cmake_cmd.append('-DMATCHMAKER_ROOT=' + matchmaker_root)
 
     if q:
         cmake_cmd.append('-Dq=ON')
@@ -184,9 +202,9 @@ def build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, ins
         cmake_cmd.append('-DCMAKE_C_COMPILER=' + clang_c)
         cmake_cmd.append('-DCMAKE_CXX_COMPILER=' + clang_cxx)
 
-    cmake_cmd.append('-Dmatchable_DIR=' + matchmaker_root + '/matchable/install' + suffix + \
+    cmake_cmd.append('-Dmatchable_DIR=' + matchmaker_root + '/matchable/install_mm' + suffix + \
             '/lib/matchable/cmake')
-    cmake_cmd.append(matchmaker_root)
+    cmake_cmd.append(stage_0_workspace_dir)
 
     if subprocess.run(cmake_cmd).returncode != 0:
         print('cmake failed')
@@ -206,22 +224,40 @@ def build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, ins
     # stage 0 finished, begin stage 1...
     os.chdir(matchmaker_root)
 
-    stage_1_matchables_dir = matchmaker_root + 'generated_include' + suffix                                \
-            + '/matchmaker/generated_matchables_stage_1'
+    stage_1_workspace_dir = build_dir + 'workspace/'
 
-    if not retain or not os.path.exists(stage_1_matchables_dir):
+    if not retain or not os.path.exists(stage_1_workspace_dir):
 
-        if not retain_matchables or not os.path.exists(stage_1_matchables_dir):
-            shutil.rmtree(stage_1_matchables_dir, ignore_errors=True)
+        # recreate stage 1 workspace
+        if not retain_leaves:
+            shutil.rmtree(stage_1_workspace_dir, ignore_errors=True)
+        if not os.path.exists(stage_1_workspace_dir):
+            os.makedirs(stage_1_workspace_dir)
+            shutil.copy(matchmaker_root + 'workspace_common/CMakeLists.txt',
+                        stage_1_workspace_dir + 'CMakeLists.txt')
+            shutil.copy(matchmaker_root + 'workspace_common/config.cmake.in',
+                        stage_1_workspace_dir + 'config.cmake.in')
+
+            os.symlink(stage_0_workspace_dir + 'generated_src', stage_1_workspace_dir + 'generated_src')
+            os.makedirs(stage_1_workspace_dir + 'generated_include/matchmaker')
+            os.symlink(stage_0_workspace_dir + 'generated_include/matchmaker/generated_letters',
+                       stage_1_workspace_dir + 'generated_include/matchmaker/generated_letters')
+
+        if (not retain_matchables and not retain_leaves) or not os.path.exists(stage_1_workspace_dir       \
+                + 'generated_include/matchmaker/generated_matchables'):
+            shutil.rmtree(stage_1_workspace_dir + 'generated_include/matchmaker/generated_matchables',
+                          ignore_errors=True)
+            shutil.copytree(stage_0_workspace_dir + 'generated_include/matchmaker/generated_matchables',
+                            stage_1_workspace_dir + 'generated_include/matchmaker/generated_matchables')
 
             print('building data reader for stage 1 ...\n\n')
 
             # build and install stage 1 data reader
             build_data_reader_cmd = ['data_reader/stage_1/scripts/build_and_install.py']
             build_data_reader_cmd.append('-b')
-            build_data_reader_cmd.append('build' + suffix)
+            build_data_reader_cmd.append('build_mm' + suffix)
             build_data_reader_cmd.append('-i')
-            build_data_reader_cmd.append('../install' + suffix)
+            build_data_reader_cmd.append('../install_mm' + suffix)
             build_data_reader_cmd.append('-m')
             build_data_reader_cmd.append('../../matchable/install' + suffix)
             build_data_reader_cmd.append('-s')
@@ -236,16 +272,11 @@ def build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, ins
             print('\n\nreading stage 1 data ...\n\n')
 
             # run stage 1 data reader to create matchables for stage 1
-            dr_stage_1_install_dir = 'data_reader/stage_1/install' + suffix
+            dr_stage_1_install_dir = 'data_reader/stage_1/install_mm' + suffix
             dr_stage_1_binary = dr_stage_1_install_dir + '/bin/data_reader_stage_1'
             dr_stage_1_data = dr_stage_1_install_dir + '/share/matchmaker/data_reader_stage_1/data'
-            stage_0_matchables_dir = matchmaker_root + 'generated_include' + suffix                        \
-                    + '/matchmaker/generated_matchables_stage_0/'
-            stage_1_longest_word_file = matchmaker_root + 'generated_include' + suffix                     \
-                    + '/matchmaker/longest_word_stage_1.h'
 
-            run_dr_stage_1_cmd = [dr_stage_1_binary, dr_stage_1_data, stage_0_matchables_dir,
-                                  stage_1_longest_word_file]
+            run_dr_stage_1_cmd = [dr_stage_1_binary, dr_stage_1_data, stage_1_workspace_dir]
             if subprocess.run(run_dr_stage_1_cmd).returncode != 0:
                 print('data_reader_stage_1 failed')
                 exit(1)
@@ -256,12 +287,12 @@ def build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, ins
 
     cmake_cmd = ['cmake', '-DCMAKE_INSTALL_PREFIX=' + install_dir]
 
-    cmake_cmd.append('-Dstage=1')
-
     if debug:
         cmake_cmd.append('-DCMAKE_BUILD_TYPE=Debug')
     else:
         cmake_cmd.append('-DCMAKE_BUILD_TYPE=Release')
+
+    cmake_cmd.append('-DMATCHMAKER_ROOT=' + matchmaker_root)
 
     if q:
         cmake_cmd.append('-Dq=ON')
@@ -277,9 +308,9 @@ def build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, ins
         cmake_cmd.append('-DCMAKE_C_COMPILER=' + clang_c)
         cmake_cmd.append('-DCMAKE_CXX_COMPILER=' + clang_cxx)
 
-    cmake_cmd.append('-Dmatchable_DIR=' + matchmaker_root + '/matchable/install' + suffix + \
+    cmake_cmd.append('-Dmatchable_DIR=' + matchmaker_root + '/matchable/install_mm' + suffix + \
             '/lib/matchable/cmake')
-    cmake_cmd.append(matchmaker_root)
+    cmake_cmd.append(stage_1_workspace_dir)
 
     if subprocess.run(cmake_cmd).returncode != 0:
         print('cmake failed')
@@ -292,16 +323,15 @@ def build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, ins
         exit(1)
 
 
-
     os.chdir(start_dir)
 
 
 
 def main():
     try:
-        opts, args = getopt.getopt(sys.argv[1:], 'hb:i:crmj:aql:d',
-                ['help', 'build_dir', 'install_dir', 'clang', 'retain', 'retain_matchables', 'jobs',
-                'atomic_libs', 'q', 'memory_usage_limit', 'debug'])
+        opts, args = getopt.getopt(sys.argv[1:], 'hb:i:crlmj:aql:d',
+                ['help', 'build_dir', 'install_dir', 'clang', 'retain', 'retain_leaves',
+                 'retain_matchables', 'jobs', 'atomic_libs', 'q', 'memory_usage_limit', 'debug'])
     except getopt.GetoptError as err:
         print(err)
         usage()
@@ -309,6 +339,7 @@ def main():
 
     use_clang = False
     retain = False
+    retain_leaves = False
     retain_matchables = False
     jobs = str(multiprocessing.cpu_count())
     build_dir = ''
@@ -325,6 +356,8 @@ def main():
             use_clang = True
         elif o in ('-r', '--retain'):
             retain = True
+        elif o in ('-l', '--retain_leaves'):
+            retain_leaves = True
         elif o in ('-m', '--retain_matchables'):
             retain_matchables = True
         elif o in ('-j', '--jobs'):
@@ -342,8 +375,8 @@ def main():
         else:
             assert False, "unhandled option"
 
-    build_and_install(use_clang, retain, retain_matchables, jobs, build_dir, install_dir, atomic_libs,
-                      q, debug)
+    build_and_install(use_clang, retain, retain_leaves, retain_matchables, jobs, build_dir, install_dir,
+                      atomic_libs, q, debug)
 
     exit(0)
